+++ src/boards/SK-iM880A/rtc-board.c
@@ -217,7 +217,6 @@ void RtcInit( void )
 
         RtcHandle.Instance = RTC;
         RtcHandle.Init.HourFormat = RTC_HOURFORMAT_24;
-
         RtcHandle.Init.AsynchPrediv = 3;
         RtcHandle.Init.SynchPrediv = 3;
 
@@ -248,7 +247,7 @@ void RtcInit( void )
     }
 }
 
-void RtcSetSysTime( uint32_t seconds, uint32_t subSeconds )
+void RtcSetSysTime( uint32_t seconds, uint16_t subSeconds )
 {
     #warning "Functionality not supported by this platform MCU. For ClassB usage please use BeaconTimingReq instead of DeviceTimeReq."
 }
